+++ src/main/java/com/nakaligoba/backend/submit/domain/SubmitRepository.java
@@ -1,0 +1,6 @@
+package com.nakaligoba.backend.submit.domain;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface SubmitRepository extends JpaRepository<Submit, Long> {
+}
